requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
cryptography==44.0.0
Django==5.1.4
django-axes==7.0.1
django-elasticsearch-dsl==8.0
django-fernet-fields==0.6
django-formtools==2.5.1
django-otp==1.5.4
django-phonenumber-field==8.0.0
django-redis==5.4.0
django-two-factor-auth==1.17.0
elastic-transport==8.15.1
elasticsearch==8.16.0
elasticsearch-dsl==8.16.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
qrcode==7.4.2
redis==5.2.1
six==1.17.0
sqlparse==0.5.2
typing_extensions==4.12.2
urllib3==2.2.3
